refactor: unwrapToken.ts
